+++ scripts/backup.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
-# shellcheck disable=SC2006
-DB_NAME=`date +"%d-%m-%Y"`
-exec pg_dump -h localhost -U postgres db1 > "<dirPath>/$DB_NAME.sql"
+DB_DIR=$(date +"%d-%m-%Y")
+mkdir "/media/rbotDATA/RED2021/$DB_DIR"
+cd "/media/rbotDATA/RED2021/$DB_DIR"
+exec mongodump